Comment out some problematic code with filterbar
The code creating an input box was malfunctioning for some reason. I'm commenting it out for the moment. Next I'll create two different browser versions of the userscript - one will be the stable version, and the other will be the beta version, once I start working on the filterbar again.