feat: Edit filter params; update task priorities by days overdue
This commit adds two features.

Ability to edit filter params:
When the user clicks a filter, the filter's parameters are placed in the url. This function takes those parameters and places them in the search box, to allow the user to edit the parameters easily.

Update task priorities by days overdue:
The user may have included some text in the task title that describes an escalating series of priorities for the task if the task is increasingly late. For example, the task may be priority 3, but it should become priority 2 if it's more than 2 weeks late, and priority 1 if it's more than 4 weeks late. This function detects the syntax used to indicate that, and updates the task's class to make it appear the correct priority. Unfortunately, this is a brittle change, since Todoist doesn't know about that change, and many user actions will cause Todoist to wipe out these changes. So I'll need to have this function run frequently.