feat: Add default priority gradients to repeating tasks
If I haven't written an explicit task gradient like [10:1] in the task title, I'm now able to infer a task gradient based on the task's period (if it's a repeating task). I upgrade it one priority for each period it's late. So a weekly task gets bumped up in priority for each week it's overdue.