fix: Convert from css priority scale to UI scale
The UI uses a 1-4 priority scale (with 1 as the highest priority), while the css classes use a 4-1 scale (4 is highest). When I retrieved the task's current priority, I was forgetting to convert it to the 1-4 scale. This commit fixes that.

Another fix: it fixes the issue I ran into before of the script overlooking tasks due "Yesterday" because I was searching for "yesterday" (all lowercase).